rename DRF's test app to drftestapp to avoid confusion
