+++ tests/contrib/djangorestframework/conftest.py
@@ -61,10 +61,10 @@ def pytest_configure(config):
                 "django.contrib.redirects",
                 "django.contrib.contenttypes",
                 "elasticapm.contrib.django",
-                "tests.contrib.djangorestframework.testapp",
+                "tests.contrib.djangorestframework.drftestapp",
                 "rest_framework",
             ],
-            ROOT_URLCONF="tests.contrib.djangorestframework.testapp.urls",
+            ROOT_URLCONF="tests.contrib.djangorestframework.drftestapp.urls",
             DEBUG=False,
             SITE_ID=1,
             TEMPLATE_DEBUG=False,
@@ -72,7 +72,7 @@ def pytest_configure(config):
             ELASTIC_APM={
                 "METRICS_INTERVAL": "0ms",
                 "TRANSPORT_CLASS": "tests.fixtures.DummyTransport",
-                "SERVICE_NAME": "testapp",
+                "SERVICE_NAME": "drftestapp",
                 "CENTRAL_CONFIG": False,
                 "CLOUD_PROVIDER": False,
                 "DJANGO_TRANSACTION_NAME_FROM_ROUTE": True,
