+++ tests/contrib/djangorestframework/drftestapp/__init__.py
