+++ tests/contrib/djangorestframework/drftestapp/views.py
